#1 Add folder Models, format code, add docs PHP
