+++ backend/app/Models/Admin.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Foundation\Auth\User as Authenticatable;
+
+class Admin extends Authenticatable
+{
+    /**
+     * The table associated with the model.
+     *
+     * @var string
+     */
+    protected $table = 'admins';
+
+    /**
+     * The table associated with the model.
+     *
+     * @var string
+     */
+    protected $fillable = [
+        'username',
+        'password',
+        'email',
+    ];
+
+    /**
+     * The attributes that should be hidden for serialization.
+     *
+     * @var string[]
+     */
+    protected $hidden = ['password'];
+
+    /**
+     * Indicates if the model should be timestamped.
+     *
+     * @var bool
+     */
+    public $timestamps = false;
+}
