--- backend/app/Student.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Student extends Model
-{
-    protected $table = 'students';
-
-    protected $fillable = [
-        'name',
-        'birthday',
-    ];
-
-    public $timestamps = false;
-}
